[web] Move ProposalPage options to its own file
